v9

+++ prj.codeforces/0276c.cpp
@@ -41,7 +41,7 @@ int main() {
 
     std::sort(mas.rbegin(), mas.rend());
     std::sort(cnt_mas.rbegin(), cnt_mas.rend());
-    long long answer = 0ll;
+    int64_t answer = 0ll;
 
     for (int i = 0; i < n; i++) {
         answer += mas[i]  * 1ll * cnt_mas[i];
